requirements.txt
absl-py==0.11.0
aiocontextvars==0.2.2
aiohttp==4.0.0a1
alembic==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
BentoML==0.9.2+19.ge743eea.dirty
bleach==3.2.1
boto3==1.16.16
botocore==1.19.16
cachetools==4.1.1
Cerberus==1.3.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.1
contextvars==2.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.3+cu102
docker==4.3.1
entrypoints==0.3
Flask==1.1.2
future==0.18.2
fvcore==0.1.2.post20201111
google-auth==1.23.0
google-auth-oauthlib==0.4.2
greenlet==0.4.17
grpcio==1.27.2
gunicorn==20.0.4
humanfriendly==8.2
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.14
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.0a1
jedi==0.17.2
Jinja2==3.0.0a1
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
loguru==0.5.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==2.0.0a1
matplotlib==3.3.2
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
ply==3.11
portalocker==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0rc2
psutil==5.7.3
ptyprocess==0.6.0
py==1.9.0
py-zipkin==0.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.1
Pygments==2.7.2
pyparsing==3.0.0b1
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.7
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
thriftpy2==0.4.12
toml==0.10.2
torch==1.5.0+cu92
torchvision==0.6.0+cu92
tornado==6.1
tqdm==4.51.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
yarl==1.6.2
zipp==3.4.0
